% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-data-frame.R
\name{as.data.frame.gdns_response}
\alias{as.data.frame.gdns_response}
\title{Coerce a gdns query response answer to a data frame}
\usage{
\method{as.data.frame}{gdns_response}(x, ...)
}
\arguments{
\item{x}{a `gdns_response` object}

\item{...}{unused}
}
\description{
Helper function to get to the `Answer` quickly
}
