% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdns.r
\name{bulk_query}
\alias{bulk_query}
\title{Vectorized query, returning only answers in a data frame}
\usage{
bulk_query(hosts)
}
\arguments{
\item{hosts}{character vector of hosts to query}
}
\value{
\code{data.frame} of only answers (use \code{query()} for detailed responses)
}
\description{
Vectorized query, returning only answers in a data frame
}
\examples{
hosts <- c("rud.is", "dds.ec", "r-project.org", "rstudio.com", "apple.com")
gdns::bulk_query(hosts)
}
\references{
\url{https://developers.google.com/speed/public-dns/docs/dns-over-https}
}

