% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dns_classes}
\alias{dns_classes}
\title{DNS CLASSes (dataset)}
\format{
data frame with columns: \code{decimal}, \code{hexadecimal}, \code{name}, \code{reference}
}
\usage{
data('dns_classes')
}
\description{
DNS CLASSes
}
\note{
As noted in , Multicast DNS can only carry DNS records with classes in the range 0-32767.
Classes in the range 32768 to 65535 are incompatible with Multicast DNS.

Last updated 2019-06-27 11:16:48
}
\references{
\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-2}

\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml}

rfc6895
}
\keyword{datasets}
