% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dns_glob_names}
\alias{dns_glob_names}
\title{Underscored and Globally Scoped DNS Node Names (dataset)}
\format{data frame with columns: \code{rr_type}, \code{node_name}, \code{reference}}
\usage{
data('dns_glob_names')
}
\description{
Underscored and Globally Scoped DNS Node Names
}
\note{
Last updated 2019-06-27 11:16:48
}
\references{
\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#underscored-globally-scoped-dns-node-names}

\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml}

rfc8552
}
\keyword{datasets}
