% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dns_opcodes}
\alias{dns_opcodes}
\title{DNS OpCodes (dataset)}
\format{
data frame with columns: \code{op_code}, \code{name}, \code{reference}
}
\usage{
data('dns_opcodes')
}
\description{
DNS OpCodes
}
\note{
Last updated 2019-06-27 11:16:48
}
\references{
\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-5}

\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml}

rfc6895, rfc1035
}
\keyword{datasets}
