% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dns_rcodes}
\alias{dns_rcodes}
\title{DNS RCODEs (dataset)}
\format{
data frame with columns: \code{rcode}, \code{name}, \code{description}, \code{reference}
}
\usage{
data('dns_rcodes')
}
\description{
DNS RCODEs
}
\note{
Last updated 2019-06-27 11:16:48
}
\references{
\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-6}

\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml}

rfc6895, rfc1035
}
\keyword{datasets}
