% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{edns0_option_codes}
\alias{edns0_option_codes}
\title{DNS EDNS0 Option Codes (OPT) (dataset)}
\format{data frame with columns: \code{value}, \code{name}, \code{status}, \code{reference}}
\usage{
data('edns0_option_codes')
}
\description{
DNS EDNS0 Option Codes (OPT)
}
\note{
Registrations made by standards-track documents are listed as "Standard," and by non-standards-track documents as "Optional." Registrations for which there are no final specifications are listed as "On-Hold."

Last updated 2019-06-27 11:16:48
}
\references{
\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-11}

\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml}

rfc6891, 3604
}
\keyword{datasets}
