% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spf.r
\name{has_spf}
\alias{has_spf}
\title{Test for whether a DNS TXT record is an SPF record}
\usage{
has_spf(spf_rec)
}
\arguments{
\item{spf_rec}{a character vector of DNS TXT records}
}
\description{
Test for whether a DNS TXT record is an SPF record
}
