% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spf.r
\name{is_soft_fail}
\alias{is_hard_fail}
\alias{is_soft_fail}
\alias{passes_all}
\title{SPF "all" type test}
\usage{
is_soft_fail(spf_rec)

is_hard_fail(spf_rec)

passes_all(spf_rec)
}
\arguments{
\item{spf_rec}{a character vector of DNS TXT records}
}
\description{
SPF "all" type test
}

