% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rrtypes}
\alias{rrtypes}
\title{Resource Record (RR) TYPEs (dataset)}
\format{data frame with columns: \code{type}, \code{value}, \code{meaning}, \code{reference}, \code{template}, \code{registration_date}}
\usage{
data('rrtypes')
}
\description{
Resource Record (RR) TYPEs
}
\note{
Last updated 2019-06-27 11:16:48
}
\references{
\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml#dns-parameters-4}

\url{https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml}

rfc6895, rfc1035
}
\keyword{datasets}
