% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spf.r
\name{is_soft_fail}
\alias{is_soft_fail}
\alias{is_hard_fail}
\alias{passes_all}
\title{SPF "all" type test}
\usage{
is_soft_fail(spf_rec)

is_hard_fail(spf_rec)

passes_all(spf_rec)
}
\arguments{
\item{spf_rec}{a character vector of DNS TXT records}
}
\value{
logical
}
\description{
SPF "all" type test
}
\examples{
is_soft_fail("v=spf1 include:_spf.apple.com include:_spf-txn.apple.com ~all")
is_hard_fail("v=spf1 include:_spf.apple.com include:_spf-txn.apple.com ~all")
passes_all("v=spf1 include:_spf.apple.com include:_spf-txn.apple.com ~all")
}
