% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spf.r
\name{spf_ipv4s}
\alias{spf_exists}
\alias{spf_includes}
\alias{spf_ipv4s}
\alias{spf_ipv6s}
\alias{spf_ptrs}
\title{SPF field extraction functions}
\usage{
spf_ipv4s(spf_rec)

spf_ipv6s(spf_rec)

spf_includes(spf_rec)

spf_ptrs(spf_rec)

spf_exists(spf_rec)
}
\arguments{
\item{spf_rec}{a character vector of DNS TXT records}
}
\description{
Various helper functions to extract SPF record components.
}

