% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spf.r
\name{split_spf}
\alias{split_spf}
\title{Split out all SPF records in a domain's TXT record}
\usage{
split_spf(spf_rec)
}
\arguments{
\item{spf_rec}{a character vector of DNS TXT records}
}
\description{
Given a vector of TXT records, this function will return a list of vectors
of all the SPF records for each. If the given TXT record is not an SPF
record, \code{NULL} is returned (which makes it easy to skip with \code{purrr}
functions).
}

